+++ package.json
@@ -39,6 +39,8 @@
     "@typescript-eslint/parser": "^4.31.0",
     "ava": "^3.13.0",
     "cross-env": "^7.0.2",
+    "eslint": "^7.32.0",
+    "eslint-import-resolver-typescript": "^2.4.0",
     "execa": "^5.0.0",
     "prettier": "^2.1.2"
   },
